Add a dtmf checker app
This app transmits a sequence of dtmf tones and verifies they are
received on the far end. Make note that before playing tones the caller
is instructed to first play a stream of silence.
Add `models.Session.send_dtmf` to support immediate tone transmission.

Additionally, when loading an `event_callback` for an event type for which
there is no `handler` defined, automatically add the
`EventListener.lookup_sess` as a default. This requires that
`marks.get_callbacks` always yields `handler` marked funcs first to
guarantee that any app defined handlers are inserted beforehand.

Other fixes:
- do not define session and job maps in `EventListener.__init__`
  declaration since they will shared amongst all instances...
- show obj mem address in `con_repr`