Measure app must count all sessions in the cluster
In order to properly capture the entire session count when measuring
results from a cluster, a `SlavePool` count method should be invoked
and stored in the metrics array. The pool can optionally be passed in
at app load time. If not provided the app's currently registered
`EventListener` will be used instead.