Add call rate tracking to metrics app
Use the session create time event time stamp to calculate instantaneous
call rates. This adds `inst_rate` and `wm20_rate` attrs to the
`CallMetrics` wrapper and graphs them in a subplot. The latter is
a rolling average rate over a window of 20 calls. Note that to compute
rate info from time stamps the metrics array must first be sorted by the
time index. This compensates for out of order metrics insertions - a side
effect of using multiple threads.

Additionally,
- increase the precision for all the collected metrics
- simplify the plot helper func such that no `fieldspec` preprocessing is
  required