Numpy is a soft dependency as well
If numpy is not installed then simply drop metrics collection altogether.

This also partially fixes #3 without implementing any solution for
listener state resynchronization after connection failures.
There doesn't seem to be an easy way to handle this afaik without
doing some janky channel checking with 'show channels' and trying to
clear/set state on the client side. I think it's safe to assume that if
the connection drops the user is responsible for investigation and state
clean up for now.