Sort a view when calc-ing call rates
Add a `view` attr which creates a new wrapper around a numpy
new-from-template view into the underlying buffer up to the current
insertion index. A ref to this view should be used for calculating
call rates and handed to the plotting routine for graphing.
Previously a one-off view created in `__getattr__` was being
mutated by methods like `np.sort` yielding a no-op for any
such mutating operation.

- bump the moving avg call rates to use a window of length 100
- update docstrings