Make `PlayRec` able to simulate conversations
This extra functionality allows the `switchy.apps.players.PlayRec` app
to be configured to play an audio clip alternating between the caller
and callee so as to simulate a conversation. Options have been added to
specify the number of times the signal should be played on each channel
as well as the "record rate" which specifies how many calls should be
recorded out of the total processed. An optional callback can also be
provided which will be invoked for each recorded call and will be passed
the paths to the transmitted (caller's) and received (callee's) recording
files.

Additional changes were required to accommodate this new feature:
- add a `Session.breakapp` method which acts as an interface to `uuid_break`
- add a `vars` sub-namespace, `caller` attr and a `get_peer` helper
  method to `Call` instances

Also, augment `Session.playback` to support playing multiple files, playback
specific parameters, and optional endless playback functionality.