`PlayRec` learns stereo recording
Now optionally takes a `rec_stereo` option to enable.
Re-key the references in the `call2rec` dict to be caller/callee
which is a bit more clear when referencing stereo recordings.
Pass a single named tuple to to the provided callback thus avoiding
potential api issues if more recordings are taken per call in the
future.

Additional changes:
- `build_originate_cmd` now inserts additional **kwargs directly
  into the {<parameters>} prefix.
- Alias `Client.host` to `.server`