Make `PlayRec` load settings aware
By knowing the default speech clip's length we can dynamically adjust the
number of times we play it (in simulating a conversation) to match the
expected duration of a call during a load test. This requires that the
driving originator does not auto-hangup calls instead leaving this to
`PlayRec` such that hangup only happens once all playback/recording
is complete.

Record calls around the rate at which they are offered such that we get
approximately 1 recording + callback per second.

Also,
- add a channel create callback which marks all calls not to be auto hung up
- make most logging debug level to avoid overwhelming output during load
  testing