Allow for `prepost` kwargs passing when `Originator` loads apps
Often it is required that an app is configured data other then the
defaults. Allow the `apps` kwarg to the `Originator` to be either
a sequence of apps or a sequence of tuples of the form `(app,
kwargs_dict)` where `kwargs_dict` is a simple dictionary specifying the
kwargs which will be provided to the `prepost` method of the app at load
time.