Port `Metrics` app to new `models.Call` api
This makes latency measurement capture explicit with regard to session
referencing within a call. Add a new column to the array which tracks
the latency from originate-to-create event time. Drop kwargs from
`__init__` since `Client.load_app` doesn't support it and thus it
won't work when controlling a cluster (eg. when using an `Originator`).