Make `Originator` sched hangups in bj handler
Since we know originates are issued using `bgapi` there should always
be a job corresponding to every call. Receiving the bg job for such
sessions means the call has connected and the onus is on us for
scheduling hangup. This avoids bothering the slave's scheduler with
a hangup if the call never connected in the first place and
additionally allows apps to mark calls which should not be
auto-hungup in a "CHANNEL_ANSWER" callback (since it comes before
the "BACKGROUND_JOB" event).

Additionally,
- drop the random call duration stuff; it makes more sense to
  compute it from a variable call rate so we don't break the
  erlang relation