Add an @app decorator to annotate switchy apps
This allows for collection of built-in apps in a global registry
which can be managed using helpers in the `switchy.apps` module.
@app can accept a `name` argument which will be used as the key for
registration. The help info will be taken from the namespace's doc string.
Modules imported for their apps can fail with import errors given that
the failure is due to an optional missing dependency (eg. numpy).