Add off-line data set loading and plotting support
Support the same off-line data retrieval api as the old numpy
implementation. Since data sets are now stored on disk in an HDF file,
we must use some other serialization format to store figure layout
information. For now, pickling the figure specification data structure is
the easiest option and also allows us to store the HDF file path
adjacently. This means that two files are written when
`Measurers.to_store` is called - an .hdf5 and a .pkl file.

- Maintain support for loading legacy numpy structured array data
  sets by first converting to a preprocessed pandas data frame.
- port the cli `plot` sub-command to match
- fix a bug in `DataStorer.buflen` such that the single row sized
  buffer case is handled correctly.