Don't leak call model references
This fixes a call model leaking issue when tracked sessions can't be
associated into a call using the current `Listener.call_id_var` (which
currently happens when the DUT doesn't forward x-headers). Adjust handler
logic to always allocate `Call`s for unknown/untracked sessions. Any
session that is hung up but can not be mapped to a known call should
therefore be reported as an error. Add a unit test to match.