Move away from HDF5
Add an abstraction layer (duck-typed interface) for storage 'backends'
This should make it easy for the addition of other backends (store types)
in the future. Add the `CSVStore` and make it the default storage format
(that's right in csv files). Keep support for an HDF5 backend but don't
make it the default (and definitely don't recommend it). Update all the
unit tests (including parameterizations for both backends).

Closes #23