Support composite app loading with `sync_caller`
Dictionary keys are of course a set so a composite app
must be passed as a list of tuples. Handle this in the
`observe.active_client` as well as app unloading.
Fix up `sync_caller` to get the job result with an error
on timeout and use the new `EventListener.app_id_vars`
list api for app lookups.