Support Python3.5 without pandas
Thanks to Simon Gomizelj (vodik) for doing the initial bit of work on
this. It actually turned out to be easier then expected with only
a few subtle changes needed regarding iterator usage; the rest was mostly
import syntax.

This change set shows that 3.5 is supported according to the unit
test suite without the using the pandas/measurement subsystem. Since the
tests aren't yet exhaustive there may be some small issues that will surface
as we start using py3 in production. There will be no support < 3.5 other
then 2.7.

The issues with pandas are related to handling unicode vs byte strings
and need to be looked at in depth.

Notable related changes include:
- Updating tox.ini to use a generative envlist covering both python versions.
- Adding `pytest-runner` as a dep for using setuptools to do test invocation.