Add support for specifying a dest-url and gateway
Allow the user to specify a --gateway name to use for originating calls
instead of a SIP profile. Adds the `gateway` kwargs to the originate command builder
func such that if provided it overrides the `profile` arg. Also adds a --dest-url
flag to allow setting the full "dial string" from the cli.

- use the `Originator.load_app()` api
- add a --loglevel flag

Resolves #40