PatternCaller improvements
`call_matches()` now always matches against a value looked up in a
caller provided mapping. This makes it easy to check against `Session`
event data. When registering a pattern-function pair you can now specify
the 'field' (like in XML dialplan conditions but in this case the name
of and event header) to match against; when not provided a default channel
variable is used.