Slight routing rework
Turn the `PatternCaller` into a `PatterRegistrar` which simply delivers
matching route functions as partials via a generator method `iter_matches`
and move it into the routing apps module. This allows for error handling
directly in the `Router` app as opposed to dealing with it inside the
pattern matcher. Use the `Router.StopRouting` exception to signal
'break-on-true' behaviour (thanks @moises-silva). Reject calls with the
more appropriate "NO_ROUTE_DESTINATION" (SIP 404) response.