Add a unified connection API
Introduces a drop-in replacement for the SWIG connection type but
wrapping the new ``asyncio`` protocol. The API for the most part
can be used synchronously for use cases (i.e. the client API) where
the connection is used outside an event loop. Add a small suite of
tests which verify the API as well as the protocol-parser using
some static event stream data files.