Wait up to 5 secs for hupalls to clear
When hupall-ing sometimes it can take a while for all calls to
tear down. Wait up to 5 seconds before error. Also, pass through the
client connection to `Job` futures. This allows for logging warnings
about jobs for which a response/ack event wasn't received fast enough
(can happen when the auto-dialer is still bursting during a hupall)
instead of raising a timeout error.