Coroutine wrapping is unnecessary unless debugging
It seems that unless running an event loop in debug mode futures can in
fact be scheduled from another thread without issue since all the
underlying machinery seems to handle loop referencing correctly.

For now turn off debug mode and add a FIXME to remind us that the
stdlib might need a patch (in `BaseEventLoop.call_soon()`).

Thanks to @vodik for poking at the initial reasoning for
`coroutinize()` which is removed here.