Don't cancel tasks in hangup handler
Doing this in the handler results in cancellation of coroutines which
have not yet had a chance to be sent a "CHANNEL_HANGUP" event due to
handlers being invoked first in the event loop. Instead, use the `done()`
method on models to determine if their lifetime has expired. Move coro
resuming before callbacks as they are previously spawned tasks whereas
callbacks are newly spawned.