Use `CHANNEL_DESTROY` as session destructor
The `CHANNEL_DESTROY` event is transmitted after a channel has been
fully torn down (Similar to `CHANNEL_HANGUP_COMPLETE`) and includes
cdr related stats that are more useful to our `CDR` app. This allows
coroutines to wait on post `CHANNEL_HANGUP` events (such as events
that are fired when a dialplan app is killed - eg. `PLAYBACK_STOP`)
prior to waiting on this final event that indicates teardown.

- Move the `CDR` app to use `CHANNEL_DESTROY` for stats collection
- Make `CHANNEL_HANGUP` use the `lookup_sess` handler

Resolves #23